+++ src/m_config.c
@@ -78,8 +78,8 @@ extern float            gp_deadzone_left;
 extern float            gp_deadzone_right;
 extern int              gp_sensitivity;
 extern dboolean         gp_swapthumbsticks;
-extern dboolean         gp_vibrate_damage;
-extern dboolean         gp_vibrate_weapons;
+extern int              gp_vibrate_damage;
+extern int              gp_vibrate_weapons;
 extern char             *iwadfolder;
 extern dboolean         messages;
 extern float            m_acceleration;
@@ -239,8 +239,8 @@ static default_t cvars[] =
     CONFIG_VARIABLE_FLOAT_PERCENT(gp_deadzone_right,                                 NOALIAS    ),
     CONFIG_VARIABLE_INT          (gp_sensitivity,                                    NOALIAS    ),
     CONFIG_VARIABLE_INT          (gp_swapthumbsticks,                                BOOLALIAS  ),
-    CONFIG_VARIABLE_INT          (gp_vibrate_damage,                                 BOOLALIAS  ),
-    CONFIG_VARIABLE_INT          (gp_vibrate_weapons,                                BOOLALIAS  ),
+    CONFIG_VARIABLE_INT_PERCENT  (gp_vibrate_damage,                                 NOALIAS    ),
+    CONFIG_VARIABLE_INT_PERCENT  (gp_vibrate_weapons,                                NOALIAS    ),
     CONFIG_VARIABLE_STRING       (iwadfolder,                                        NOALIAS    ),
     CONFIG_VARIABLE_FLOAT        (m_acceleration,                                    NOALIAS    ),
     CONFIG_VARIABLE_INT          (m_doubleclick_use,                                 BOOLALIAS  ),
@@ -675,11 +675,9 @@ static void M_CheckCVARs(void)
     if (gp_swapthumbsticks != false && gp_swapthumbsticks != true)
         gp_swapthumbsticks = gp_swapthumbsticks_default;
 
-    if (gp_vibrate_damage != false && gp_vibrate_damage != true)
-        gp_vibrate_damage = gp_vibrate_damage_default;
+    gp_vibrate_damage = BETWEEN(gp_vibrate_damage_min, gp_vibrate_damage, gp_vibrate_damage_max);
 
-    if (gp_vibrate_weapons != false && gp_vibrate_weapons != true)
-        gp_vibrate_weapons = gp_vibrate_weapons_default;
+    gp_vibrate_weapons = BETWEEN(gp_vibrate_weapons_min, gp_vibrate_damage, gp_vibrate_weapons_max);
 
     if (m_doubleclick_use != false && m_doubleclick_use != true)
         m_doubleclick_use = m_doubleclick_use_default;