+++ src/m_config.h
@@ -172,9 +172,13 @@ typedef enum
 
 #define gp_swapthumbsticks_default              false
 
-#define gp_vibrate_damage_default               true
+#define gp_vibrate_damage_min                   0
+#define gp_vibrate_damage_default               100
+#define gp_vibrate_damage_max                   100
 
-#define gp_vibrate_weapons_default              true
+#define gp_vibrate_weapons_min                  0
+#define gp_vibrate_weapons_default              100
+#define gp_vibrate_weapons_max                  100
 
 #define health_min                              0
 #define health_default                          100