+++ src/p_pspr.c
@@ -216,7 +216,7 @@ void P_FireWeapon(player_t *player)
 
     if (gp_vibrate_weapons && vibrate)
     {
-        int     motorspeed = weaponinfo[readyweapon].motorspeed;
+        int     motorspeed = weaponinfo[readyweapon].motorspeed * gp_vibrate_weapons / 100;
 
         if ((readyweapon == wp_fist && player->powers[pw_strength])
             || (readyweapon == wp_chainsaw && linetarget))
@@ -275,11 +275,11 @@ void A_WeaponReady(mobj_t *actor, player_t *player, pspdef_t *psp)
     //  if player is dead, put the weapon away
     if (pendingweapon != wp_nochange || !player->health)
     {
-        if ((gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
+        if (gp_vibrate_weapons && vibrate)
         {
             if (pendingweapon == wp_chainsaw)
             {
-                idlemotorspeed = CHAINSAWIDLEMOTORSPEED;
+                idlemotorspeed = CHAINSAWIDLEMOTORSPEED * gp_vibrate_weapons / 100;
                 XInputVibration(idlemotorspeed);
             }
             else if (idlemotorspeed)